Identifiers without a common base class for now

+++ share/tasks.py
@@ -132,13 +132,14 @@ class MakeJsonPatches(celery.Task):
             started_by = ShareUser.objects.get(pk=started_by_id)
         logger.info('%s started make JSON patches for %s at %s', started_by, normalized, datetime.datetime.utcnow().isoformat())
 
+        cs = None
         try:
             cs = ChangeSet.objects.from_graph(ChangeGraph.from_jsonld(normalized.normalized_data, extra_namespace=normalized.source.username), normalized.id)
             if cs and (normalized.source.is_robot or normalized.source.is_trusted):
                 # TODO: verify change set is not overwriting user created object
                 cs.accept()
         except Exception as e:
-            if cs.id:
+            if cs and cs.id:
                 cs.delete()
             raise self.retry(countdown=10, exc=e)
